Speed up and cache get_settings() result
The get_settings() command is called constantly. Each call it was
gathering information on the entire callstack, which is a very slow
operation. This patch changes the get_settings() call to only query
the relevant callstack frames, and then (on the assumption that its
always supposed to return the same result) caches the result. This
signficantly speeds up nearly every operation (since get_settings()
is called so often).