working to fix issue #237
scaling of mesh was not taken into account when gedge is from mesh.

also, the inner gverts are moved to behave better when gedge is from mesh.
note: this is not a great solution.  polystrips now assumes that if the outer
gverts of a gedge are from the mesh, and that there is a quad strip between
the two gverts, then the gedge *should* be from the mesh.  to work around
this, the artist needs to explicitly add an extra gvert between the two gverts.