"fix" for issue #382
for some reason, a vertex has NaN in its coordinate, which causes
problems.

upon startup, RF will now call `validate()` on the mesh data, which
_should_ correct these issues (setting co to 0).

i added an assertion in `RFVert.co.setter` to detect setting `co` to
NaN, which will cause assertion failure, but it will also point out
where the NaNs are coming from.

because of the last sentence, i call this a "fix" rather than a proper
fix.