code clean up, bug fixes, optimized
cleaning:

- removed tons of old, usused code
- renamed some vars and fns to be clearer
- moved self.update_tool_options call to rftool.py

bugs:

- tried tracking down strange bug. i believe the bug has something to do
with caching, but not certain. touching the undo stack seems to flush
the cache to circumvent the problem, but this is a workaround.  to
reproduce:
    - start PS, select strip, drag but cancel moving a handle, exit RF
    - start ps again, drag (already visible) handle... strip doesn't
move
- undoing would only partly switch back to the previous tool.  now it
works correctly.  (TODO: should we switch tools upon undo?  switching
helps put the edit into context, but changes the working context!)

optimize:

- added selection version, so caches don't dirty with selection