fixed issue #403
the offset on the max distance a ray can travel before hitting a source
mesh (for visibility check) now depends on the bbox of the sources.  the
larger the source, the greater the offset.

this is not ideal, but it seems to work in the cases that i've tested.
non ideal because i believe it boils down to a floating point errors,
and this is just a workaround.