fix for issue #432 (hack)
seems that some graphics cards will report "Vertex shader was
successfully compiled to run on hardware." when compiling a shader has
succeeded.  i had (incorrectly) assumed that any report was an
indication of failure.  added a quick test for the presence of 'was
successfully compiled' in the log.  my previous check was a hack; this,
too, is a hack.  need to rework this.