blender 2.91.0alpha removed a property (issue #921)
blender 2.91.0alpha removed the
`bpy.context.preferences.input.wheel_scroll_lines` property, which was
"currently hidden and unused".  this change in blender caused RF to
crash at init.

changed internal scroll amount to a fixed `3`, which is the same as what
the default was and is what Blender is currently using.