start of split target visualization
whenever a tool starts manipulating a relatively small portion of the
target, the visualization is split into static and dynamic
visualizations, and only the dynamic viz is updated.  once manipulation
is committed or canceled, the split is cleared.  this viz split allows
for much faster feedback for artist when the target gets large.