fix issue #1022
when artist presses alt+tab to switch to different app then returns to
blender, the ALT is still considered pressed.  added a set of actions
(`reset_actions`) that will cause the current state of actions to be
reset.

note: a problem still exists (was not created by this commit), where a
pressed button is not recognized when returning to blender, as these
keys were not pressed when Blender had focus.  without being able to
poll the current state, i'm not sure how to work around this issue
anyways.  (can poll modifiers, ex:
https://docs.blender.org/api/current/bpy.types.Event.html#bpy.types.Event.alt,
but not arbitrary keys)