major updater bug fix!
the updater will only overwrite certain files during update by default
(to prevent overwriting settings, for example), and we have to specify
the extensions that we allow to be overwritten.  the problem was that
.html files (basically all UI) was not in that list.  from 3.2.0 to
3.2.1, we had changed the name of one of the options and we updated the
UI accordingly.  However, because the .html extension was not in the
overwrite list, the updater did not overwrite the 3.2.0 html ui files,
which contained the previous option name.  so, once blender started, and
the UI system tried looking up the value, it broke but it wasn't
reported to the console.  instead, RF/CC kept going on with the
initialization, which caused a cascade of errors since the UI wasn’t
fully set up.

this commit adds .html to the overwrite list.

the following changes are also in this commit:
- for consistency, the LICENSE files have been renamed to have .LICENSE
extension
- addon updater code is updated to the latest version on the Blender
Add-on Updater repo on GitHub
- a few other extensions were added to overwrite and pre-delete lists