fix for issue #1042
on some macs (newer M1), the graphics driver (and therefore blender)
would crash hard when compiling a shader that has a `discard` statement
inside a `switch`.  this commit works around this issue by moving the
`discard` statements into (rooughly) equivalent `if` conditions.

a couple of notes:

- this might also fix issues #1025, #915, #879, #753.  in those issues,
the driver didn't crash, but there were definite visual artifacts
related only to the UI shader.

- the only switched-discard statements were in the UI shader, which is
likely why the visual artifacts in issues of previous note only showed
up with drawing UI and _NOT_ with any of the gizmos.