fix for issue #1074
if any modifier is pressed or unpressed while Blender is in an operator (ex: changing view), RetopoFlow's view of that modifier becomes out of sync.  The only workaround was to press that modifier once back in RF.  this change will always look to the current state of all modifiers, so it will always be in correct sync state.

we do lose the ability to distinguish between left and right modifiers (ctrl, shift, alt), because this isn't exposed through bpy.  however, we were not using these anywhere, so it's not a deal breaking change for RF.