added option to keep orbit center when nothing is selected (#1182)
before, when nothing was selected, the orbit center is place at the
center of bbox of all sources.  this changes the behavior of orbitting
when `Orbit Around Selection` is enabled in Edit > Preferences >
Navigation.  unchecking this option will make orbitting in RF similar to
that in Blender, but with the loss of the ability to center view on all
sources.