fix for issue #1135
I _think_ what was happening is this:  at certain blender ui scales, the
innertext of summary would be computed at a slightly larger (or equal?)
width to the details container, and therefore it would get wrapped but
the height would not get increased.  this displayed as not having text.
reducing the padding on right and bottom worked around this issue (in
all my tests).

Really, the UI system needs rewritten (on long-term to-do)