Updated general.md
The View Options section includes the word "think" instead of "thick," which makes no sense in the context of describing the size of edges. So, the correct version of the phrase would be:
"The Target Drawing options control the rendering of the target mesh. The Above and Below options control transparency of the target mesh. Vertex Size and Edge Size control how large the vertices and how thick the edges are."