updated left/right mouse selection code (PR #1064, commit 73c7cfc)
using keyconfigs.user rather than keyconfigs.active with fallback to
keyconfigs.default, because the user KeyConfig has all of these combined
in the way that Blender does it.

see
https://docs.blender.org/api/current/bpy.types.KeyConfigurations.html#bpy.types.KeyConfigurations.user

removed the old way of getting mouse select option (Blender < 2.90?).

RF now reloads the keymaps on startup so that changes to mouse selection
option are seen by restarting RF (not needing to restart Blender).