better detection of when an op takes control
in order for RF to detect when an external operator takes control, we
are tracking the mouse position (hack).  if ever the mouse position is
goes out of sync (in other words, another operator runs modal and
doesn't pass through the mouse), we can assume that RF does not have
control anymore.  the problem is if RFOperator does not pass through,
the RF thinks RF lost control, even though it hasn't.  Now, RFOperator
will update RFCore to improve this hack.