workaround for issue #1281
it seems Blender issue T93410
(https://projects.blender.org/blender/blender/issues/93410) is an issue
with Blender 4.0+ again.

this commit works around that issue by not drawing dropdown when CC is
running.  (this could be problematic if RF crashes and the CC variable
doesn't get reset, but restarting Blender will correct this)