fix issue #1321
`context.scene.objects` may include items that are not in the active
view layer, which causes `ray_cast` to fail because it has no eval'd
mesh data (do not need to eval it if it's not visible).  switched to
`context.view_layer.objects` to only iterate over objects in the view
layer.