Removed RetopoFlow from tool names
I liked the idea of adding RetopoFlow to the tool names in the toolbar, but it made the name way too long and the actual tool name was getting cut off when the toolbar was expanded to a normal width. We could use 'RF' instead, but I like the idea that we are owning THE Strokes tool or THE Contours tool.