polystrips: vertical slice through #1455
right now, only considering bmverts that share a face with a selected
bmvert.

right now, selected bmverts are moved with a factor of 1.0, and these
un-selected neighboring bmverts are moved with a constant factor of 0.5.

later, we can determine factor as a falloff based on distance from curve
and size of brush (or some property of edit)