workaround for issue #1489
the Screencast Keys add-on was always moving to top of modal stack (it
needs to do this so that it can see what actions the artist performs in
order to display them).

However, we were looking at the top spot to see if RF is the most active
(as opposed to starting the knife tool, for example).

so, created a hard-coded workaround specifically for the Screencast Keys
add-on.

need to see if there is a better approach for this at some point.