possible workaround for issue #1515
this commit will run periodic tests to see if there are any depsgraph
pre/post callbacks that could invalidate the edit mesh bmesh.  we will
keep track of any that do in a list.

then, when a RF operator starts, all these methods get temporarily
removed from the callback list.

when the RF operator finishes, they get re-appended to the callback
list.

i have not tested to see how much of a performance impact this has, or
what the visual effect is...