Improved Relax (#1586)
removed algorithm strength (can simply use brush strength) and added max
distance limiter options: brush radius, average edge length, and bounce
detector.

brush radius.
the distance any vert is displaced is limited to some factor of the
brush radius.  technically, this was already in the code, but it was
hard coded to 0.125.  now, the artist has control over the factor.

average edge length.
the distance any vert is displaced is limited to some factor of the
average length of edges connected to vert.  note: using min length can
cause problem if there are any 0-length edges (verts at same location)

bounce detector.
the distance a vert is displaced will decay if bounces are detected.
this detection is based on the direction of the displacement for current
iteration vs the last iteration.  if the two directions are pointing
opposite directions (>90deg between directions), then the amount a vert
is moved is halved, but this halving compounds.  in other words, if 4
bounces are detected for a vert, that vert's displacement distance is
halved 4 times (0.5 * 0.5 * 0.5 * 0.5 = 0.0625).  Note: this will mean
that bouncing verts will quickly get "stuck" and not move.  to counter
this, all decay amounts are reset each time you start relaxing.