5.0 port: Fix crash when exiting RetopoFlow
Some introduced breaking changes in Blender 5.0 API are causing `ACCESS_VIOLATION` errors when exiting RetopoFlow.

This is due to the fact that the session data is stored as a custom property on the text datablock via text["data"], which is not supported in Blender 5.0 anymore, per Blender 5.0 Python API docs: "Removal of (Unsupported) Access to Runtime-Defined Properties Storage Data".
As a side effect, it seems to make impossible to add custom data (dict type, in this case) to an ID datablock by using a custom property. In other words, that dict was treated as a different type of data handled by Blender rather than a dict type and that type of data was "unsupported" per the docs, now the little support it had is totally removed.

This commit fixes the issue by storing the session data inside the text datablock content itself, by parsing the data to a JSON format back and forth.
It also introduces a cache for the session data to avoid multiple reads from the same text datablock in runtime.