Fix shader compilation error for some devices #1664
Error message: `No matching overloaded function found: blender_srgb_to_framebuffer_space`

This commit fixes this issue by removing the `blender_srgb_to_framebuffer_space` function added by Retopoflow to the shader source code,
as is a function that is "added" to the compilation process by Blender itself.
This double definition of the same function was causing the compilation error, as the GPU finds two definitions of the same function,
and it doesn't know which one to use as both have the same signature and RF-defined one is not a valid overloaded function.
While this issue is ignored in most devices, it can cause an error in some devices, possibly due to specific drivers or backends,
making those users unable to enable and use Retopoflow.