Refactor shader color `mix` fn to use `vec4` as the interpolation value type instead of `float` (fix #1664)
This commit addresses the following shader compilation error:
```
Cannot convert from: "const float" to: "out highp 4-component vector of vec4"
```

This error seems to occur in specific devices, tho the root cause is not yet determined it could be somewhat related to a driver bug or compiler issues.
This changes should not affect the functionality of the shaders nor the rest of the existing users experience.