Contours: 98,76% time improvement + fix #1603
This commit fixes #1603 ("Contour issues on non-uniform meshes"), as well as drastically improves the tool time.
Time Improvement: 98,47%, from ~ 0.4979 seconds to ~ 0.0076 seconds in a single 1.2M tris source-mesh and a fresh target-mesh.

The original logic tried to project the new vertices onto the surface of the source meshes. However, the method it used was prone to fail when working over complex meshes, those with concave-convex shapes (such as my character's ears) and probably small-scale meshes as well as other corner cases we don't know yet. The location calculated for the raycast could make it prone to fail as the point itself could be inside the mesh, and making it to position the vertices in the other side of a concave/convex surface shape or in the other side of the mesh, making it look like if the contours ring is shrinked or twisted in some cases.

Change the way we calculate the position for each vertex.

Instead of using ray_from_point_through_point and raycast_ray_valid_sources that are raycast-based functions, with issues with the commented kind of source meshes, let's use nearest_point_valid_sources which uses the Blender built-in closest_point_on_mesh function for each source mesh, and that is not only fast but has shown as truly precise in this scenario.

Contours tool was the slowest tool in Retopoflow V4!
Part of that was the expensive processing and some redundant loops.

The solution for this issue has increased the performance.
Removing 2 redundant loops has increased the performance by an extra.